Change forex API and add forex caching
Note: This represents a major change to operation. API keys are now needed in
order to access updated forex data, but if a cache file is available the tool
should now work without internet access.

Cache file is json and uses this format:
{
  'dest_curr_1': {
    'source_curr': ###
    ...
  },
  'dest_curr_2': {
    'source_curr': ###
    ...
  }
  ...
}

JPY conversion data MUST be available - in cache mode conversions are done by
first converting to JPY, then to the final currency.